+++ src/app/components/Footer.tsx
@@ -4,11 +4,11 @@ import Link from 'next/link';
 import Image from 'next/image';
 import React from 'react';
 import styles from '../styles/Home.module.scss';
-import { NavigationLink } from './Header';
+import { NavigationLink } from './NavigationLink';
+import classNames from 'classnames';
+import { basePath } from '../utils/constants';
 
 export default function Footer() {
-  const basePath = '/dibbs-site';
-
   const [expanded, setExpanded] = React.useState(false);
   const onClick = () => {
     if (window.innerWidth < 1024) setExpanded((prvExpanded) => !prvExpanded);
@@ -17,22 +17,21 @@ export default function Footer() {
   const testItemsMenu = [
     <Link
       href="/"
-      key="one"
-      className={`usa-nav__link ${styles.homeNavItem}`}
+      className={classNames('usa-nav__link', styles.homeNavItem)}
       onClick={onClick}
     >
       <span className={styles.navbarItemText}>Home</span>
     </Link>,
-    <NavigationLink key="two" href={`/`} onClick={onClick}>
+    <NavigationLink href="/" onClick={onClick}>
       Home
     </NavigationLink>,
-    <NavigationLink key="two" href={`/our-products`} onClick={onClick}>
+    <NavigationLink href="/our-products" onClick={onClick}>
       Our products
     </NavigationLink>,
-    <NavigationLink key="two" href={`/our-products`} onClick={onClick}>
+    <NavigationLink href="/case-studies" onClick={onClick}>
       Case studies
     </NavigationLink>,
-    <NavigationLink key="three" href={`/engage-with-us`} onClick={onClick}>
+    <NavigationLink href="/engage-with-us" onClick={onClick}>
       Engage with us
     </NavigationLink>,
   ];
@@ -45,14 +44,16 @@ export default function Footer() {
             <div className="usa-logo">
               <em className="usa-logo__text">
                 <a href="http://cdc.gov" title="<Project title>">
-                  <span className={`${styles.navbarLogoText} sr-only`}>
+                  <span
+                    className={classNames('sr-only', styles.navbarLogoText)}
+                  >
                     CDC US center for disease control and prevention
                   </span>
                   <Image
                     width={200}
                     height={40}
                     alt=""
-                    className={'margin-x-0'}
+                    className="margin-x-0"
                     src={`${basePath}/images/CDC.svg`}
                   />
                 </a>
