+++ src/app/utils/constants.ts
@@ -1,0 +1,6 @@
+const basePath = '/dibbs-site';
+const flagImg = `${basePath}/images/us-flag.png`;
+const dotGovIcon = `${basePath}/images/us-gov-icon.svg`;
+const httpsIcon = `${basePath}/images/https-icon.svg`;
+
+export { basePath, flagImg, dotGovIcon, httpsIcon };
